+++ components/exercises/chapter-2-jon/JChessBoard.tsx
