+++ components/exercises/chapter-2-jon/JPyramid.test.tsx
