+++ components/exercises/chapter-2-jon/JPyramid.tsx
