+++ components/exercises/chapter-2-jon/jGetChessBoard.tsx
