+++ components/exercises/chapter-2-jon/jGetPyramid.tsx
