+++ components/exercises/chapter-3-jon/JGetMinimumNumber.tsx
