+++ components/exercises/chapter-3-jon/JMinimum.tsx
