+++ components/exercises/chapter-3-jon/jGetNumberOfChars.tsx
